make windows version in 2010

+++ src/learner/evaluation-inl.hpp
@@ -133,7 +133,7 @@ struct EvalCTest: public IEvaluator {
       }
       wsum += base_->Eval(tpred, tinfo);
     }
-    return wsum / ngroup;
+    return static_cast<float>(wsum / ngroup);
   }
 
  private:
